refactor : RacingGame 클래스 분리
