+++ src/test/java/racingcar/controller/RaceControllerTest.java
@@ -2,8 +2,7 @@ package racingcar.controller;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import racingcar.controller.RaceController;
-import racingcar.model.Car;
+import racingcar.domain.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
