+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,7 +1,6 @@
-package racingcar.model;
+package racingcar.domain;
 
 import org.junit.jupiter.api.Test;
-import racingcar.model.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
